# src/swagger.json
{
    "openapi": "3.0.0",
    "title:": "Project FullStack",
    "description": " Fornecimento de dados referente a administracao de um mini restaurante ",
    "version": "1.0.0",
    "paths": {
        "/auth/cliente/register":{
            "post":{
                "summary":"Gera um token para um cliente ",
                "description": "Rota responsavel por lidar com a geração de token",
                "tags":[
                    "Autenticação"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/authRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/authResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }

            }
        },
        "/auth/usuario/register":{
            "post":{
                "summary":"Gera um token para um cliente ",
                "description": "Rota responsavel por lidar com a geração de token",
                "tags":[
                    "Autenticação"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/authRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/authResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }

            }
        },
        "/clientes": {
            "get": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary": "Busca todos os clientes",
                "description": "Rota responsavel por retornar todos os clientes",
                "tags": [
                    "Cliente"
                ],
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary": "Adiciona um novo Cliente",
                "description": "Rota responsavel por adicionar um novo cliente",
                "tags": [
                    "Cliente"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/addCliente"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary": "Rota responsavel por atualizar um cliente",
                "description": "Rota responsavel por atualizar um cliente ja existente",
                "tags": [
                    "Cliente"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/cliente"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clientes/{idCliente}": {
            
            "get": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "parameters": [
                    {
                      "in": "path",
                      "name": "idCliente",
                      "required": true,
                      "schema": {
                        "type": "number"
                      },
                      "description": "Id do Cliente"
                    }
                  ],
                "summary": "Busca por um unico Cliente",
                "description": "Rota responsavel por retornar um unico cliente",
                "tags": [
                    "Cliente"
                ],
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/clientes/{idCliente}/pedidos/{idPedido}": {
            
            "get": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "parameters": [
                    {
                      "in": "path",
                      "name": "idCliente",
                      "required": true,
                      "schema": {
                        "type": "number"
                      },
                      "description": "Id do Cliente"
                    },
                    {
                        "in": "path",
                        "name": "idPedido",
                        "required": false,
                        "schema": {
                          "type": "number"
                        },
                        "description": "Id do Pedido"
                      }
                  ],
                "summary": "Busca por Pedido(s) do Cliente",
                "description": "Rota responsavel por retornar pedido(s) do cliente",
                "tags": [
                    "Cliente"
                ],
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pedidos": {
            "get": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary": "Busca todos os pedidos",
                "description": "Rota responsavel por retornar todos os pedidos",
                "tags": [
                    "Pedido"
                ],
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary": "Adiciona um novo Pedido",
                "description": "Rota responsavel por adicionar um novo pedido",
                "tags": [
                    "Pedido"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/addPedido"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary": "Rota responsavel por atualizar um pedido",
                "description": "Rota responsavel por atualizar um pedido ja existente",
                "tags": [
                    "Pedido"
                ], 
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/pedido"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pedidos/{idPedido}": {
            "get": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "parameters": [
                    {
                      "in": "path",
                      "name": "idPedido",
                      "required": true,
                      "schema": {
                        "type": "number"
                      },
                      "description": "Id do Pedido"
                    }
                ],
                "summary": "Busca por um unico Pedido",
                "description": "Rota responsavel por retornar um unico pedido",
                "tags": [
                    "Pedido"
                ],
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/pedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "parameters": [
                    {
                      "in": "path",
                      "name": "idPedido",
                      "required": true,
                      "schema": {
                        "type": "number"
                      },
                      "description": "Id do Pedido"
                    }
                ],
                "summary": "Exclui um Pedido",
                "description": "Rota responsavel por deletar um unico pedido",
                "tags": [
                    "Pedido"
                ],
                "responses": {
                    "Response":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                    
                }
            }
        },
        "/itemspedidos/{idPedido}": {
            "get": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "parameters": [
                    {
                      "in": "path",
                      "name": "idPedido",
                      "required": true,
                      "schema": {
                        "type": "number"
                      },
                      "description": "Id do Pedido"
                    }
                  ],
                "summary": "Busca por uma lista de itens daquele pedido",
                "description": "Rota responsavel por todos os itens daquele pedido",
                "tags": [
                    "Itens do Pedido"
                ],
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/itemPedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/itemspedidos": {
            "post": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary":"Adiciona um novo item para um pedido",
                "description":"Rota responsavel por adicionar itens em um pedido",
                "tags": [
                    "Itens do Pedido"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/addItemPedido"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/itemPedido"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/items/{idItem}": {
            "get": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "parameters": [
                    {
                      "in": "path",
                      "name": "idItem",
                      "required": true,
                      "schema": {
                        "type": "number"
                      },
                      "description": "Id do Item"
                    }
                  ],
                "summary": "Busca por um unico Item",
                "description": "Rota responsavel por retornar um unico item",
                "tags": [
                    "Item"
                ],
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/item"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "parameters": [
                    {
                      "in": "path",
                      "name": "idItem",
                      "required": true,
                      "schema": {
                        "type": "number"
                      },
                      "description": "Id do Item"
                    }
                ],
                "summary": "Exclui um Item",
                "description": "Rota responsavel por deletar um unico item",
                "tags": [
                    "Item"
                ],
                "responses": {
                    "Response":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                    
                }
            }
        },
        "/items": {
            "get": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary": "Busca por todos os itens",
                "description": "Rota responsavel por retornar todos os itens",
                "tags": [
                    "Item"
                ],
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/item"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary":"Adiciona um novo item",
                "description":"Rota responsavel por adicionar um item",
                "tags": [
                    "Item"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/addItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/item"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                       "bearerAuth": []
                    }
                 ],
                "summary": "Rota responsavel por atualizar um item",
                "description": "Rota responsavel por atualizar um item ja existente",
                "tags": [
                    "Item"
                ], 
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#components/schemas/item"
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/item"
                                }
                            }
                        }
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#components/schemas/status"
                                }
                            }
                        }
                    }
                }
            }
        }
        
    },
    "components": {
        "schemas": {
            "authRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "senha": {
                        "type": "string"
                    }
                }
            },
            "authResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "number"
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
            "addCliente": {
                "type": "object",
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "cpf": {
                        "type": "string"
                    },
                    "telefone": {
                        "type": "string"
                    }
                }
            },
            "cliente": {
                "type": "object",
                "properties": {
                    "idCliente": {
                        "type": "number"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "cpf": {
                        "type": "string"
                    },
                    "telefone": {
                        "type": "string"
                    }
                }
            },
            "addPedido":{
                "type":"object",
                "properties":{
                    "descricao": {
                        "type": "string"
                    },
                    "fkCliente": {
                        "type":"number"
                    },
                    "dataPedido": {
                        "type": "string"
                    }
                }       
            },
            "pedido":{
                "type":"object",
                "properties":{
                    "idPedido": {
                        "type":"number"
                    },
                    "descricao": {
                        "type": "string"
                    },
                    "fkCliente": {
                        "type":"number"
                    },
                    "dataPedido": {
                        "type": "string"
                    }
                }       
            },
            "addItemPedido": {
                "type": "object",
                "properties":{
                "fkPedido": {
                    "type":"number"
                },
                "fkItem": {
                    "type":"number"
                },
                "quantidade": {
                    "type":"number"
                },
                "preco_unitario": {
                    "type":"string"
                }}
            },
            "item": {
                "type":"object",
                "properties":{
                    "idItem": {
                        "type":"number"
                    },
                    "nome": {
                        "type":"string"
                    },
                    "preco_unitario": {
                        "type":"string"
                    },
                    "imagem_64": {
                        "type":"string"
                    }
                }
            },
            "addItem": {
                "type":"object",
                "properties":{
                    "nome": {
                        "type":"string"
                    },
                    "preco_unitario": {
                        "type":"string"
                    },
                    "imagem_64": {
                        "type":"string"
                    }
                }
            },
            "itemPedido": {
                "type":"object",
                "properties":{
                "id": {
                    "type":"number"
                },
                "fkPedido": {
                    "type":"number"
                },
                "fkItem": {
                    "type":"number"
                },
                "quantidade": {
                    "type":"number"
                },
                "preco_unitario": {
                    "type":"string"
                }}
            },
            "status": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes":{
            "bearerAuth":{
                "type": "http",
                "scheme": "Bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": {
        "bearerAuth": ["bearerAuth"]
    }
}